Use checkbox instead of images for boolean fields.
