+++ src/app/protected/cranix/security/firewall/firewall.component.html
@@ -41,6 +41,10 @@
           </ion-segment>
         </ion-col>
         <ion-col offset="3" size="3" class="ion-align-self-center ion-text-end hidden-xs hidden-sm hidden-md">
+          <ion-button *ngIf="segment == 'in'" size="small" (click)="applyIncomingRules()"
+            matTooltip="{{'Add outgoing rule.' | translate }}">
+            <ion-icon slot="icon-only" name="checkmark-sharp"></ion-icon>
+          </ion-button>
           <ion-button *ngIf="segment == 'out'" size="small" (click)="addOutgoinRule()"
             matTooltip="{{'Add outgoing rule.' | translate }}">
             <ion-icon slot="icon-only" name="add-circle-outline"></ion-icon>
@@ -59,100 +63,93 @@
       </ion-row>
     </ion-card-header>
     <ion-card-content>
-      <form #incomingForm="ngForm" (ngSubmit)="applyIncomingRules(incomingForm.value)">
-        <ion-grid *ngIf="segment == 'in'">
-          <ion-row>
-            <ion-col size=2>
-              <ion-item>
-                <ion-label>{{ 'ssh access' | translate }}</ion-label>
-              </ion-item>
-            </ion-col>
-            <ion-col size=1>
-              <ion-item lines="none">
-                <ion-toggle slot="start" color="secondary" checked="{{incoming.ssh}}" [(ngModel)]="incoming.ssh"
-                  name="ssh">
-                </ion-toggle>
-              </ion-item>
-            </ion-col>
-            <ion-col offset=6 size=1>
-              <ion-button type="submit" class="ion-no-margin" matTooltip="{{'Apply changes.' | translate }}">
-                <ion-icon slot="icon-only" color="success" name="checkmark-sharp"></ion-icon>
-              </ion-button>
-            </ion-col>
-          </ion-row>
-          <ion-row>
-            <ion-col size=2>
-              <ion-item>
-                <ion-label>{{'admin access' | translate}}</ion-label>
-              </ion-item>
-            </ion-col>
-            <ion-col size=1>
-              <ion-item lines="none">
-                <ion-toggle slot="start" color="secondary" checked="{{incoming.admin}}" [(ngModel)]="incoming.admin"
-                  name="admin">
-                </ion-toggle>
-              </ion-item>
-            </ion-col>
-          </ion-row>
-          <ion-row>
-            <ion-col size=2>
-              <ion-item>
-                <ion-label>{{'https access' | translate}}</ion-label>
-              </ion-item>
-            </ion-col>
-            <ion-col size=1>
-              <ion-item lines="none">
-                <ion-toggle slot="start" color="secondary" checked="{{incoming.https}}" [(ngModel)]="incoming.https"
-                  name="https">
-                </ion-toggle>
-              </ion-item>
-            </ion-col>
-          </ion-row>
-          <ion-row>
-            <ion-col size=2>
-              <ion-item>
-                <ion-label>{{'rdesktop access' | translate}}</ion-label>
-              </ion-item>
-            </ion-col>
-            <ion-col size=1>
-              <ion-item lines="none">
-                <ion-toggle slot="start" color="secondary" checked="{{incoming.rdesktop}}"
-                  [(ngModel)]="incoming.rdesktop" name="rdesktop">
-                </ion-toggle>
-              </ion-item>
-            </ion-col>
-          </ion-row>
-          <ion-row>
-            <ion-col size=2>
-              <ion-item>
-                <ion-label>{{'open other ports' | translate}}</ion-label>
-              </ion-item>
-            </ion-col>
-            <ion-col size=10>
-              <ion-item>
-                <ion-input type="text" value="{{incoming.other}}" [(ngModel)]="incoming.other" name="other">
-                </ion-input>
-              </ion-item>
-            </ion-col>
-          </ion-row>
-        </ion-grid>
-      </form>
-      <ag-grid-angular *ngIf="segment == 'out'" 
+      <ion-grid *ngIf="segment == 'in'">
+        <ion-row>
+          <ion-col size=2>
+            <ion-item>
+              <ion-label>{{ 'ssh access' | translate }}</ion-label>
+            </ion-item>
+          </ion-col>
+          <ion-col size=2>
+            <ion-item lines="none">
+              <ion-toggle slot="start" color="secondary" checked="{{incoming.ssh}}" [(ngModel)]="incoming.ssh"
+                name="ssh">
+              </ion-toggle>
+            </ion-item>
+          </ion-col>
+        </ion-row>
+        <ion-row>
+          <ion-col size=2>
+            <ion-item>
+              <ion-label>{{'admin access' | translate}}</ion-label>
+            </ion-item>
+          </ion-col>
+          <ion-col size=2>
+            <ion-item lines="none">
+              <ion-toggle slot="start" color="secondary" checked="{{incoming.admin}}" [(ngModel)]="incoming.admin"
+                name="admin">
+              </ion-toggle>
+            </ion-item>
+          </ion-col>
+        </ion-row>
+        <ion-row>
+          <ion-col size=2>
+            <ion-item>
+              <ion-label>{{'https access' | translate}}</ion-label>
+            </ion-item>
+          </ion-col>
+          <ion-col size=2>
+            <ion-item lines="none">
+              <ion-toggle slot="start" color="secondary" checked="{{incoming.https}}" [(ngModel)]="incoming.https"
+                name="https">
+              </ion-toggle>
+            </ion-item>
+          </ion-col>
+        </ion-row>
+        <ion-row>
+          <ion-col size=2>
+            <ion-item>
+              <ion-label>{{'rdesktop access' | translate}}</ion-label>
+            </ion-item>
+          </ion-col>
+          <ion-col size=10>
+            <ion-item lines="none">
+              <ion-toggle slot="start" color="secondary" checked="{{incoming.rdesktop}}"
+                [(ngModel)]="incoming.rdesktop" name="rdesktop">
+              </ion-toggle>
+            </ion-item>
+          </ion-col>
+        </ion-row>
+        <ion-row>
+          <ion-col size=2>
+            <ion-item>
+              <ion-label>{{'open other ports' | translate}}</ion-label>
+            </ion-item>
+          </ion-col>
+          <ion-col size=10>
+            <ion-item>
+              <ion-input type="text" value="{{incoming.other}}" [(ngModel)]="incoming.other" name="other">
+              </ion-input>
+            </ion-item>
+          </ion-col>
+        </ion-row>
+      </ion-grid>
+      <ag-grid-angular *ngIf="segment == 'out'"
         style="width: 100%; height: 100%;"
         class="{{authService.settings.agGridThema}}"
         [rowData]="outData"
-        [gridOptions]="outOptions" 
+        [gridOptions]="outOptions"
         [context]="context"
         rowSelection="multiple"
         (gridReady)="outGridReady($event)"
         (selectionChanged)="outSelectionChanged()"
         id="outGridTable">
       </ag-grid-angular>
-      <ag-grid-angular *ngIf="segment == 'remote'" 
+      <ag-grid-angular *ngIf="segment == 'remote'"
         style="width: 100%; height: 100%;"
         class="{{authService.settings.agGridThema}}"
         [rowData]="remoteData"
-        [gridOptions]="remoteOptions" 
+        [gridOptions]="remoteOptions"
         [context]="context"
         rowSelection="multiple"
         (gridReady)="remoteGridReady($event)"
