+++ src/app/shared/cranix-shared.module.ts
@@ -26,7 +26,7 @@ import { PipesModule } from 'src/app/pipes/pipe-modules';
 import { AgGridModule } from 'ag-grid-angular';
 import { ActionBTNRenderer } from 'src/app/pipes/ag-action-renderer';
 import { YesNoBTNRenderer } from 'src/app/pipes/ag-yesno-renderer';
-import { YesNoTextBTNRenderer } from 'src/app/pipes/ag-yesno-text-renderer';
+import { CheckBoxBTNRenderer } from 'src/app/pipes/ag-checkbox-renderer';
 import { GroupActionBTNRenderer } from 'src/app/pipes/ag-group-renderer';
 import { UserActionBTNRenderer } from 'src/app/pipes/ag-user-renderer';
 import { DateCellRenderer } from 'src/app/pipes/ag-date-renderer';
@@ -67,7 +67,7 @@ import { UserIdToNameCellRenderer } from 'src/app/pipes/ag-userid-to-name-render
     UserIdToNameCellRenderer,
     ToolbarComponent,
     YesNoBTNRenderer,
-    YesNoTextBTNRenderer
+    CheckBoxBTNRenderer
   ],
   imports: [
     AgChartsAngularModule,
@@ -92,7 +92,7 @@ import { UserIdToNameCellRenderer } from 'src/app/pipes/ag-userid-to-name-render
     UserIdCellRenderer,
     UserIdToNameCellRenderer,
     YesNoBTNRenderer,
-    YesNoTextBTNRenderer,
+    CheckBoxBTNRenderer,
   ]),
   FormsModule,
   IonicModule,
